+++ apps/frontend/src/libs/components/protected-route/libs/constants/routes-without-auth-wrapper.constant.ts
@@ -1,0 +1,6 @@
+import { AppRoute } from "~/libs/enums/enums.js";
+import { type ValueOf } from "~/libs/types/types.js";
+
+const ROUTES_WITHOUT_AUTH_WRAPPER: ValueOf<typeof AppRoute>[] = [AppRoute.QUIZ];
+
+export { ROUTES_WITHOUT_AUTH_WRAPPER };