+++ packages/shared/src/modules/onboarding/libs/types/ondoarding-question-request-dto.ts
@@ -1,0 +1,9 @@
+import { type OnboardingAnswerDto } from "./onboarding-answer-dto.type.js";
+
+type OnboardingQuestionRequestDto = {
+	answers: OnboardingAnswerDto[];
+	id: number;
+	label: string;
+};
+
+export { type OnboardingQuestionRequestDto };